[dvsim] Fix PEP8 error and slightly tidy code in testplan_utils.py
The PEP8 violation is comparing types for equality in this line:

         elif (type(list1[key]) == type(list2[key])):

Python tends to go for duck typing, so this approach isn't
particularly Pythonic in general, but I've tried to keep the same
logic for now at least.

Since fixing this needed a bit of surgery on the code anyway, I've
used 'continue' to avoid a level of indentation and added some
comments explaining how the logic works.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>