[dvsim] Tidy up config file loading in FlowCfg.py
This changes how command line overrides work, but I think the previous
behaviour probably wasn't intentional.

Before this patch, a command line override would only take effect when
reading a second config file that set a given key where both config
files set it with a 'non-default' value. This behaviour was added by
commit 680f7e2 and I think the point was just to allow the user to set
the tool (VCS or xcelium) on the command line.

After this patch, a command line override takes effect as soon as we
see the corresponding key in any configuration file. However, we can't
allow any old command line argument. For example, there is a --waves
command line argument which clashes with a config file key.
(Presumably, the existing code would get confused if we had multiple
config files setting the waves key inconsistently, but no-one does
this). To avoid this confusion, we now list the keys that can be
overridden on the command line in _CMDLINE_FIELDS.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>