[dvsim] Fix printing of Deploy objects
The __repr__ method is supposed to give a hunk of Python code that
evaluates to something the same as the object (whereas __str__ can do
whatever you fancy). Overriding __repr__ like this caused an infinite
recursion in verbose mode if you tried to print an object (presumably
the pprint library tries to print 'self' if you give it
'self.__dict__').

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>