[dvsim] Update `tests` key behavior in regressions
This was encountered in #3040. The `common_sim_cfg.hjson` adds a
`sanity` regression target without specifying the `tests` key, with the
assumption that each individual IP's cfg hjson will update the `sanity`
regression with the required list of tests. In PR #3040, the sanity
tests in some of the IPs had to be commented out due to some other
issues. But on doing so, running `sanity` regression ended up adding ALL
available tests for regression for that particular DUT, since the `tests`
key remained empty list (originally intended behavior). This PR fixes
that behavior. So to sum up:

Previous behavior of the `tests` key was as follows:
- `tests` member of class `Regression` in `Modes.py` defaulted to empty
  list
- If in regressions (in the hjson), it continues to remain an empty
  list after all hjson source are parsed, then ALL available tests are
  run

New behavior is as follows:
- `tests` member of class `Regression` in `Modes.py` now defaults to
  `None`
- If in regressions (in the hjson) it is set to an empty list, and
  eventually remains an empty list after all hjson sources are parsed,
  then no tests are run
- If in regressions (in the hjson) it is entirely omitted, such that
  after all hjson sources are parsed, it continues to hold the value of
  `None`, the ALL available tests are run

Signed-off-by: Srikrishna Iyer <sriyer@google.com>