[dvsim] Allow testplan to be omitted
In the current implementation, `dvsim` makes it necessary to provide a
testplan, otherwise it results in errors being thrown.

This patch fixes that and allows testplan to be completely omitted. The
tests in that case appear as 'Unmapped Tests' in the final results
table. This is useful for DUTs such as `prim_lfsr` and `prim_present`.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>