[dvsim] Make the scheduling logic per-target
Because we don't run jobs for multiple targets at the same time, we
can make lots of simplifying assumptions. Some of them were in the
code already (for example, we knew that we only had to look at
previous targets to find results of jobs we depended on).

This commit takes things further, explicitly running the schedulers
for separate targets one after another.This means that the scheduler
gets much simpler (no dependencies, plus this gets rid of a layer of
looping in the run function).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>