[DVSim] Fix how sw_images is treated
This change fixes the way sw_images is treated - it can now be set in
`build_modes`, `run_modes`, test specifications and 'bare' in the HJson
which will get it applied to all tests. Previously, it was not possible
to add it to build_modes. Also a bug prevented it from getting appended
properly across run_modes and test specifications.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>