[dvsim] Statically display jobs' status
Currently, dvsim prints a new status on the terminal every 10 seconds
(or if there is a change in pass/failed/killed count). If the regression
runs for several hours, that results in a lot of status prints.

In this change, the third party progress bar manager enlighten is used
to print the status instead. We skip the progress bar to keep things
lightweight.

Based on whether output redirection has occurred or not, the Scheduler
switches between printing using enlighten v/s printing using logging
(which is how stats are printed currently).

Signed-off-by: Srikrishna Iyer <sriyer@google.com>