[dvsim] Spot fixes for LSF and internal launcher
Over the past few weeks, the scheduling and launching mechanism were
refactored to fully support LSF and our Google cloud which is used
internally.

The fixes made in this PR are minor ones to support our internal cloud
launcher (which will not be committed to our open repo).

The fixes are as follows:
- Updates to commands and env vars in the HJson data
- Addition of `input_dirs` and `output_dirs` variables in `Deploy`
  - This is needed only for the cloud launcher because jobs with cloud
  run in a hermetically sealed VM which needs to explicitly indicate
  what the input and output data is.
  - We will refactor this again in future when we will flip the Deploy
  and Launcher hierarchy.
- Fixes to command sub-strings (ensure they are quoted properly)
- Allow remote launchers to set their own python virtualenv region
- Some cosmetic changes

Signed-off-by: Srikrishna Iyer <sriyer@google.com>