[dvsim] Keep dependencies list
This set of changes is aimed at retaining the dependency order across
targets, even if the dependency is not scheduled to be run. The
Deploy::dependencies list once constructed, remains untouched. We
instead change the way the FlowCfg::deploy list is created - if targets
are not required to run (for example, when --build-only or --run-only
switch is passed), then they are not added to the deploy list. This
means that the deploy list needs to be constructed correctly - the
scheduler relies on it to know what to run.

The scheduler previously recursively went through the dependencies to
determine what all needed to be run - this required the FlowCfg to
delete dependencies after the fact, if flow modifier switches were
passed. This is no longer needed because we now explicitly provide it
the list of things to run instead. This also means when checking an
item's eligibility to be enqueued based on its dependencies' statuses,
it needs to ignore the deps that were not a part of the original deploy
list.

The reason for this change is our internal Google Cloud based launching
system, which runs each job (input -> process -> output) in an isolated
VM instance. The job's input and output are tarballs that flit between
the user's workstation, Google Cloud storage, and the VM instance. To
support --run-only for example, in our environment, the run deploy
object needs to be able to provide a pointer to its build dependency
(which would have run in the past) so that the built simulation
executable can be tarballed and uploaded as the run-job's input.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>