[dv/dvsim] Adds failure bucketizer for triage.
Changes _post_finish to take an ErrorMessage as second argument.
Moves class SimCfg's Results to its own file and renames it to SimResults.
The bucketizer is called when initializing SimResults.
This will change when reruns are implemented, since we would want to
rerun one test per bucket as soon as the failing test completes.

Signed-off-by: Guillermo Maturana <maturana@google.com>