[dvsim] Testplanner refactor & covergroup support
This PR significantly refactors the testplanner code, with the goal of
adding support for writing the coverage model as a part of the testplan.
DV users can now write covergroups planned as a part of the effort
within the testplan HJSon itself like so:
```hjson
  covergroups: [
    {
      name: foo_cg
      desc: '''Desription of covergroup foo_cg.'''
    }
  ]
```

The following changes are made:

- dvsim/testplanner/class_defs.py is moved to dvsim/Testplan.py
- Testplan.py adds a base class `TestplanElement` and derived class
`Covergroup`. The original `TestplanItem` is renamed to `Testpoint`
which extends from `TestplanElement`.
- Covergroups and testpoints are abstractions of each item in the
testplan.
- The original code is significantly overhauled in the hope to make it
more pythonic with sensible APIs the external tools can invoke.
- New features:
  - Compute testplan progress: this makes a list of all planned tests
  (i.e. the `tests` list in each testpoint) and marks them off by
  looking at the simulation results. The generated table shows the
  total number of tests to be written, and the number of tests written,
  with a progress percentage. If coverage is dumped, it will show the
  number of covergroups written too based on the simulation results
  (coming soon!).
Signed-off-by: Srikrishna Iyer <sriyer@google.com>