[dvsim] Enabling glob-style patterns for -i switch
This commit enables the user to pass glob-style patterns instead of
fully resolved test/regression names to `-i|--items` switch.

The following are some examples:
- Run all CSR tests: `-i *csr*`
- Run all sleep tests: `-i chip_sw_sleep*`

As before, it matches the input patterns with the list of available
regrssions first. For example, `-i *smoke*` will add all tests enabled
for the smoke regressions AND then, it will add all tests containing the
pattern `*smoke*` in the name that are not added to the smoke
regression. If multiple regressions match the input patterns, and there
is an overlap of tests, then the subsequent regressions are dropped with
a warning message. If an input pattern does not match any available
regression or test name, then a warning is thrown.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>