[lint] Unify lint parser scripts
This splits out common subfunctions used for parsing lint reports into a
class named LintParser.
The structure of the lint results is also slightly changed in
preparation for the next commit that adapts the Dvsim side accordingly.

Note that this is only an intermediate solution.
The parsing mechanism will be moved into the Dvsim core code at some
point (see #9079).

Signed-off-by: Michael Schaffner <msf@opentitan.org>