[dvsim/lint] Make message reporting more flexible
This extends the LintCfg class and makes the lint message reporting
mechanism more flexible, so that the lint message categories can be
tailored to the tool instead of having to rely on fixed message
categories. I.e., the categories, corresponding severities and report
labels can now be configured as part of the Dvsim Hjson configuration,
instead of being hadcoded inside the LintCfg class.

To this end, a set of helper classes called MsgBucket and MsgBuckets
is added that allow for convenient aggregation and printout of lint
tool messages.

This will allow us to reuse the lint class for upcoming tool flows like
CDC where the lint tool has different message categories than the
existing RTL lint tools.

Signed-off-by: Michael Schaffner <msf@opentitan.org>