[cdc] Add support for initial CDC flow with open-source views
This reuses the updated LintCfg flow from Dvsim and introduces a DVsim
configuration for CDC runs with VerixCDC.

Although the flow reuses the LintCfg class for reporting purposes, the
actual setup in terms of run scripts is more similar to the synthesis
flow since the CDC flow relies on an SDC file and a more involved run
script (i.e., the tool is not run directly through FuseSoC as the other
lint tools are).

CDC waivers have to be specified in one common waiver TCL file at the
moment, although we are likely going to split this out into multiple
waiver files to make this more manageable.

Signed-off-by: Michael Schaffner <msf@google.com>