[dvsim] Add support for specifying primary_build_mode
A primary build mode is the build (i.e. DUT configuration
that is considered to be the 'main' one.

It is necessary to designate one to identify the coverage
database associated with the main build configuration.
VCS in particular, mandates the coverage database associated
with the main build configuration to be supplied first to
URG when merging coverage databases from different build
configurations. Other than that, it does not serve any
purpose.

If `primary_build_mode: foo` is not specified in the
hjson, it picks the globally set `build_mode` value.
If a global `build_mode` is not set in the hjson, it
is set to "default", which is the most typical usecase.
If `primary_build_mode` is set to an invalid value,
an error is thrown. It must be set to one of the existing
`build_modes`.

This fix to ensure URG is invoked correctly will apply
for all DUTs - nothing needs to be done for a particular
DUT. It is agnostic to the tool, so it will apply for
Xcelium and other tools as well.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>