[dvsim] Use new results_server methods to list cleanly
This should not be a change from the existing code, although the logic
for iterating over recent runs has got rather simpler.

Interestingly, the code was supposed to delete old results (all but
the most recent 14). Unfortunately, the "rm" command that gets
calculated doesn't actually get run! It hasn't been run since
f578e7c (in January 2020).

This commit removes the complicated logic to calculate the rm
command (that we weren't running anyway!), but it would be pretty
trivial to resurrect it properly if in future we decide it would be
useful.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>