[dv, dvsim] Update nc flag for interactive sim
This commit replaces the -I flag for nc run command with the -wl flag
for interactive jobs. The -I flag is used to forward TTY signals to the
running job. This works well for jobs launched from terminal as a sigint
on the console will stop the job without an explicit nc stop command.
However, when interactive jobs are managed using dvsim, the -I flag
based job termination does not work well as dvsim's resources for the
interactive job gets orphaned when the nc run job is terminated by the
user. Hence switching to the -wl flag and letting dvsim handle the
sigint and stop the running nc job using an explicit nc stop command.
This also unifies the abort handling for interactive and non-interactive
jobs further simplifying the code

Signed-off-by: Venkat Krishnan <venkateswarank.opentitan@gmail.com>