[refactor] fix imports given the files are now part of a package
DVSim used to be a collection of scripts, now it is a full python package.
The imports need to use module path syntax and be relative to the dvsim
package. Full paths rather than absolute paths have been used as that makes
it easier to rename and move files later with sed style import fixups.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>