[reporting] fix css file import issue
Now we are using a python package the premailer library sees the css
file as an "external file" and refuses to use it. So we now load the
file and pass the contents to premailer instead. This was chosen instead
of enabling external loading, which is disabled by default for security.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>