feat: improved config loading and project settings
Move out the project level config functions. The cli module should
probably be a thin wrapper around internal functions rather than
containing a fairly large amount of business logic.

Improved config loading with testing, and save the config to JSON file
in the scratch dir. Currently there is no tracability in terms of what
config was actually used to produce the generated results. Add
serialisation to save the config file to the root of the scratch
dir for later reference. At some point this could be used to restart or
finish off a job that failed for some reason.

Update cli and flow config classes to use the new config loader.
Move the flow config into the project config. This means the project
config contains the whole project configuration including the flow
config. This can be broken down later into smaller Pydantic objects as
the fields and types become better defined.

Add explicit relative path field to the config loading. The rel_path is
a calculated field that is critical to determining where files are
placed and found in the scratch dir and also the reports dir. At the
moment the field is not created until the flow objects are created.
Which means that the saved config doesn't have the information unless
it's overridden in the config itself.

Move the calculated field out of the flow base class init
into the config loading stage where it is saved the filesystem in the
scratch dir along with the rest of the loaded config. Which means it can
be used to navigate the scratch dir after a run.

Move overrides from flow base to config loading stage This is to ensure
that the saved project config is already resolved otherwise the
overrides mechanism would need to be implemented in any code that loads
the project config.

Without this change, the report generation code can't see the correct
rel_path field as it is overridden in the flow object after being saved
as part of the project config.

Resolve as many wildcards as possible on config load. Now that the
overrides have moved back into the config loading stage the overridden
fields are not having their wildcards resolved for some reason.

It is useful to do this anyway as we then have more resolved config
values in the saved project config

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>