feat: [git] add repo root helper and migrate to GitPython library
With the nightly regression running functionality more git interaction
is required. Rather than implementing everything from scratch using
subprocess calls and maintaining it, add the GitPython library as a
dependency and use that for the git interactions.

The repo root helper is implemented using Pathlib as the GitPython
library doesn't seem to support this functionality. It requires a path
to the repo root exactly and fails if it doesn't get a valid repo path.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>