style: [tests] disable pedantic rules for test files.
Given the nature of tests, some things are implied or
simplified to make the tests more readable. This rule
checks for boolean passed as positional argument. Which
makes sense for the main codebase, as flags passed as
positional make it difficult to know what flag is being
set. However in the tests it's common to assert against
a boolean and noqa comments would make the tests needlessly
messy. Disabling the rule globally for test files given
the prevalence of this usecase.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>