[linting] update ruff ci rules to disable new logger rule and auto fix
There is a new ruff rule that checks for using the root logger, which we do.
Using the root logger is generally a bad thing as it means when we set the log
level to debug it also sets the log level for other libraries. Generally you
would create a dedicated logger and use that instead.

Also added better comments to explain the disabled rules and rerun
`ruff check --fix` to resolve some auto fixable errors.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>