[reporting] remove unnesesery latest dir for reporting
DVSim currently creates an unessesery extra dir `latest` in each report
dir. This was so that the old mechanism that allowed overwriting
existing dirs with a new report could rename the `latest` dir to a
timestamp and create a new dir with the latest report in it.

There were several issues with this approach, mutating previous results
is a bad idea in general as it risks corrupting the previous results and
unintentionally misleading people. More importantly in this case, the
timestamps were what ever time it was at the time of replacement. Which
meant that all the links were broken in the summary report and the paths
were semi random and not deterministic. Which efectivly rendered the
previous results unreachable without guessing the exact timestamp for
each IP block report page.

The new report publishing mechanism removes all of this mutation, which
means the index pages are preserved with all the links working. The
namespacing occurs at the top level, meaning there is no reason for the
lower level dirs to have a `latest` dir as there is only ever one report
at that level.

This PR only changes the dir structure of the reports dir, the broader
scratch directory is not effected. While it would be nice to do
something similar there, it's a larger change and would need a lot more
thought to implement.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>