refactor: [config] save the config to JSON file in the scratch dir
Currently there is no tracability in terms of what config was actually used to
produce the generated results. This commit adds serialisation to save
the config file to the root of the scratch dir for later reference.

At some point this could be used to restart or finish off a job that
failed for some reason.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>