refactor: [config] resolve wildcards on load
Resolve as many wildcards as possible on config load. Now that the overrides
have moved back into the config loading stage the overridden fields are not
having their wildcards resolved for some reason.

It is useful to do this anyway as we then have more resolved config values
in the saved project config

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>