[report] remove old report dir versioning
This mechanism for replacing dirs with timestamped dirs at the point
when the individual dirs want to be replaced is causing a number of side
effects. For example rstmgr_cnsty_chk report is generated in the sub dir
of rstmgr. Given the simple logic for backing up previous report
generations, depending on the order each flow config is processed the
backup logic renames dirs generated in the same run.

This commit removes the backup logic altogether with the intention of
doing any backup at the root dir instead. This means that in the short
term reports will be overwritten, opening the door for incremental
builds. Backing up will have to be performed manually and the scratch
dir purged before running if a fresh run is required.

It's unclear at this stage what an automated backup solution would look
like? Ideally changes in the src files would be determined and new incremental
build only running the jobs required to generate the changed output artefacts.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>