feat: add deployment object dump debug feature
Debug feature to dump a JSON representation of the `Depoly` objects,
just before they are sent to the scheduler. This feature is hidden behind
an environment variable to avoid clutering the CLI interface. The full
`Deploy` object is not JSON serialisable in it's current form, so we
only serialise the attributes that the LocalLauncher uses.

To enable set `DVSIM_DEPLOY_DUMP=true`, run DVSim and there will be a
file called `deploy-[timestamp].json` generated in the scratch
directory.

This file can they be compared between runs to check for regressions in
the flow leading up to job deployment. With a fixed seed, the resulting
json files should be identical. Generating a `diff` or hash of the two
files and comparing shows this.

Using this functionality it should now be easier to confirm that
refactoring does not introduce regressions.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>