feat: make cov_db_dirs deterministic
Sort the coverage database paths, while leaving the first path in place.
This should mean that the dumped deployment objects are now
deterministic. At the moment when comparing dumps often the only
difference is the order of these paths, adding unnecessary noise and
making it more difficult to check for meaningful differences.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>