refactor: migrate from Depoy.dump to JobSpec.model_dump
Now we have a Pydantic model that represents the full data requirement
of the scheduler, we can use the builtin `JobSpec.model_dump()` instead of
the custom `Deploy.dump()`. At the moment this model has to contain a
couple of callback functions which cannot be serialised, so these
attributes are excluded.

This commit does change the format of the dumped "deployment" objects
file. So if this is being used to check for breaking changes, then
hashes need to be compared against one generated from this commit from
now on.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>