refactor: create a higher level function to generate all reports
This higher level wrapper can be used to generate all the reports with
one function. The benefit of this is that the CLI no longer needs to know
what lower level functions are required to generate the whole set of
reports. Making it easier to ensure that the admin command report
generation doesn't diverge from the in flow report generation.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>