feat: block results report HTMX
Reduce the block report to an HTML fragment that is loaded with the
wrapper using AJAX requests. Update the summary report links to use a
bootstrap button class (which is nicer than the blue URLs anyway?), and
use HTMX to replace the `#report-content` div with the block report.

The `hx-push-url="true"` is important to make sure the back button on
the browser works as expected.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>