refactor: report use the local css/js
Add a static file renderer to go alongside the jinja2 renderer for the
static dependencies that do no need templating. This essentially creates
a copy of the files in the output directory.

Now we have the css/js dependencies locally generated we can update the
report template to use them.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>